Disallow using syntax-form identifiers as variables. (#387)
At present things like `(when if ...)` are legal syntax, with the `if` interpreted as a variable reference bound to to an unusable internal value. This is quite wrong, but the behavior was used in a couple places.

The common implementation of `thunk` and `||` needs to be in a separate module because `define_syntax` bodies are compiled before `define`s are expanded, so they cannot reference helpers from the same module.  Racket has `define_for_syntax` to create helpers at the correct expansion phase.

Fixes #126